refactor: create base client class
